Removes Task 0 bug
MoveChatbotHere() gets called when the answer graph is constructed. When this function is called, ownership of the _chatBot pointer moves from ChatLogic into a GraphNode, which means moves semantics were needed. This commit supplies a patch with std::move and comments out the delete call in ChatLogic. A later commit will likely refactor so that delete responsibilities are more obvious.